Move MainFrameCounterTestImpl to test folder
