+++ cobalt/shell/renderer/shell_content_renderer_client.cc
@@ -48,7 +48,6 @@
 #include "v8/include/v8.h"
 
 #if defined(RUN_BROWSER_TESTS)
-#include "cobalt/shell/common/main_frame_counter_test_impl.h"   // nogncheck
 #include "cobalt/shell/common/shell_test_switches.h"            // nogncheck
 #include "cobalt/shell/renderer/shell_render_frame_observer.h"  // nogncheck
 #include "content/public/test/test_service.mojom.h"             // nogncheck
@@ -246,9 +245,6 @@ void ShellContentRendererClient::ExposeInterfacesToBrowser(
 #if defined(RUN_BROWSER_TESTS)
   binders->Add<mojom::TestService>(
       base::BindRepeating(&CreateRendererTestService),
-      base::SingleThreadTaskRunner::GetCurrentDefault());
-  binders->Add<mojom::MainFrameCounterTest>(
-      base::BindRepeating(&MainFrameCounterTestImpl::Bind),
       base::SingleThreadTaskRunner::GetCurrentDefault());
 #endif  // defined(RUN_BROWSER_TESTS)
   binders->Add<web_cache::mojom::WebCache>(
