+++ cobalt/testing/browser_tests/common/main_frame_counter_test.mojom
